[BUGFIX] Use correct options
Instead of using the options_heap in the loop that triggers
oc_render_start_drawing() the options_to_process variable
is used because this is the one that is processed in the
loop.

With the old version it was possible that the wrong options
are used when rendering multiple orgcharts.